+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ----------
 
 v1.0.0 (??.??.????)
+ * skipPersistedCheck check in simple marker added
  * getSortbyFieldForTable method added to get sortable column of a table
  * move up and down links in form util added to move records in a be module
  * new Wrapper Class Tx_Rnbase_Service_Base (don't forget to load)  for \TYPO3\CMS\Core\Service\AbstractService or t3lib_svbase which is compatible for all TYPO3 versions